refacto(api): move user-reconciliation-service to shared
