+++ api/src/prescription/learner-management/domain/usecases/index.js
@@ -1,11 +1,11 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import * as userReconciliationService from '../../../../../lib/domain/services/user-reconciliation-service.js';
 import * as userRepository from '../../../../identity-access-management/infrastructure/repositories/user.repository.js';
 import * as organizationFeatureApi from '../../../../organizational-entities/application/api/organization-features-api.js';
 import * as obfuscationService from '../../../../shared/domain/services/obfuscation-service.js';
 import * as placementProfileService from '../../../../shared/domain/services/placement-profile-service.js';
+import * as userReconciliationService from '../../../../shared/domain/services/user-reconciliation-service.js';
 import { logErrorWithCorrelationIds } from '../../../../shared/infrastructure/monitoring-tools.js';
 import * as libOrganizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../../shared/infrastructure/repositories/organization-repository.js';
@@ -53,7 +53,7 @@ import { importStorage } from '../../infrastructure/storage/import-storage.js';
  * @typedef {import('../../../organization-learner/infrastructure/repositories/registration-organization-learner-repository.js')} registrationOrganizationLearnerRepository
  * @typedef {import ('../../infrastructure/repositories/student-repository.js')} studentRepository
  * @typedef {import ('../../infrastructure/repositories/sup-organization-learner-repository.js')} SupOrganizationLearnerRepository
- * @typedef {import ('../../../../../lib/domain/services/user-reconciliation-service.js')} UserReconciliationService
+ * @typedef {import ('../../../../shared/domain/services/user-reconciliation-service.js')} UserReconciliationService
  * @typedef {import('../../../../identity-access-management/infrastructure/repositories/user.repository.js')} userRepository
  * @typedef {import('../../infrastructure/repositories/jobs/validate-csv-organization-learners-import-file-job-repository.js')} validateCsvOrganizationImportFileJobRepository
  * @typedef {import ('../../infrastructure/repositories/jobs/validate-organization-learners-import-file-job-repository.js')} ValidateOrganizationImportFileJobRepository
